# perf: improve tracker field filtering performance DHIS2-19910