--- a/.gitignore
+++ b/.gitignore
@@ -36,7 +36,11 @@ node_modules
 **/rebel.xml
 .vscode
 .cursor
+.claude
+CLAUDE.local.md
 # ignore generated artifact directory
 docker/artifacts
 # ignore local docker override files
 docker-compose.override.yml
+# ignore JMH benchmark results
+jmh-result*.csv