--- a/dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/export/event/EventQueryParams.java
+++ b/dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/export/event/EventQueryParams.java
@@ -379,6 +379,10 @@ class EventQueryParams {
     return accessiblePrograms != null && accessibleProgramStages != null;
   }
 
+  public Set<UID> getEnrollments() {
+    return enrollments;
+  }
+
   public EventQueryParams setEnrollments(Set<UID> enrollments) {
     this.enrollments = enrollments;
     return this;