--- a/dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/export/event/EventService.java
+++ b/dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/export/event/EventService.java
@@ -87,7 +87,8 @@ public interface EventService {
    * TrackerIdSchemeParams}.
    */
   @Nonnull
-  Event getEvent(UID uid, @Nonnull TrackerIdSchemeParams idSchemeParams, EventFields fields)
+  Event getEvent(
+      @Nonnull UID uid, @Nonnull TrackerIdSchemeParams idSchemeParams, @Nonnull EventFields fields)
       throws NotFoundException;
 
   /**