--- a/dhis-2/dhis-support/dhis-support-commons/src/main/java/org/hisp/dhis/commons/jackson/config/JacksonObjectMapperConfig.java
+++ b/dhis-2/dhis-support/dhis-support-commons/src/main/java/org/hisp/dhis/commons/jackson/config/JacksonObjectMapperConfig.java
@@ -110,7 +110,7 @@ public class JacksonObjectMapperConfig {
     return dataValueJsonMapper;
   }
 
-  @Bean("xmlMapper")
+  @Bean
   public ObjectMapper xmlMapper() {
     return xmlMapper;
   }
@@ -129,14 +129,19 @@ public class JacksonObjectMapperConfig {
   }
 
   static {
-    JtsModule jtsModule = new JtsModule(new GeometryFactory(new PrecisionModel(), 4326));
-    jtsModule.addSerializer(Geometry.class, new GeometrySerializer());
+    JtsModule jtsModule = createJtsModule();
     jsonMapper.registerModule(jtsModule);
     dataValueJsonMapper.registerModule(jtsModule);
     xmlMapper.registerModule(new JtsXmlModule());
   }
 
-  private static ObjectMapper configureMapper(ObjectMapper objectMapper) {
+  public static JtsModule createJtsModule() {
+    JtsModule jtsModule = new JtsModule(new GeometryFactory(new PrecisionModel(), 4326));
+    jtsModule.addSerializer(Geometry.class, new GeometrySerializer());
+    return jtsModule;
+  }
+
+  public static ObjectMapper configureMapper(ObjectMapper objectMapper) {
     return configureMapper(objectMapper, false);
   }