## Background

[field filtering](https://docs.dhis2.org/en/develop/using-the-api/dhis-core-version-master/metadata.html#webapi_metadata_field_filter) allows reducing the JSON response via the query parameter `fields`. The JSON response can also be changed using these [transformations](https://docs.dhis2.org/en/develop/using-the-api/dhis-core-version-master/metadata.html#webapi_field_transformers). Field filtering is used by almost every endpoint in DHIS2. Tracker uses the same implementation as platform which was [re-implemented in 2.38](https://dhis2.atlassian.net/browse/TECH-650) using Jackson. It looks to me as if analytics still depends on the [previous field filter implementation](https://github.com/dhis2/dhis2-core/tree/master/dhis-2/dhis-services/dhis-service-node/src/main/java/org/hisp/dhis/fieldfilter) at least in some endpoint(s).

I ran a [Gatling](https://github.com/dhis2/performance-tests-gatling) performance tests locally with a single user making 100 requests against `/api/organisationUnits?fields=:all,!name,!id,!favorites,!translations,!children,!sharing` sequentially. I collected the following profile

[profile.html](https://github.com/user-attachments/files/22094934/profile.html)

If you select a thread processing an HTTP request and search (Ctrl+f) for `fieldfiltering` you'll find that 90% of CPU samples are dealing with it. This gets worse quickly if you increase the `pageSize` and thus the elements to be filtered. Field filtering outways time spent in the DB even though hibernate is making lots of requests to the DB (at least one per entity)!

To get a feel for the cost of field filtering I ran the above test against

1. dhis2/core:42.0 with no code changes
2. dhis2/core-dev:42.0 where I removed the field filtering feature, so no matter `fields` we serialize the entire object using Jackson

For our default `pageSize=50` we already pay ~20ms on field filtering even though 2. is at a disadvantage as it has to serialize more than 1. as there is no filtering.

<img width="1000" height="600" alt="analysis" src="https://github.com/user-attachments/assets/e4126171-6df4-451f-868b-b1a052b53342" />

The main reasons why the current implementation is slow are

1. The implementation is built around the full path of a field:

[FieldFilterService](https://github.com/dhis2/dhis2-core/blob/2151f37440c411b844ee59aa6c6a1420012f6dcb/dhis-2/dhis-services/dhis-service-field-filtering/src/main/java/org/hisp/dhis/fieldfiltering/FieldFilterService.java#L250) turns `fields=:all,!name,!id,!favorites,!translations,!children,!sharing` into

```
paths = {ArrayList@31490}  size = 71
 0 = {FieldPath@31270} "FieldPath(name=parent, path=[], exclude=false, preset=false, transformers=[], fullPath=parent)"
 1 = {FieldPath@31415} "FieldPath(name=yearToDate, path=[queryMods], exclude=false, preset=false, transformers=[], fullPath=queryMods.yearToDate)"
 2 = {FieldPath@31272} "FieldPath(name=type, path=[], exclude=false, preset=false, transformers=[], fullPath=type)"
 3 = {FieldPath@31483} "FieldPath(name=minDate, path=[queryMods], exclude=false, preset=false, transformers=[], fullPath=queryMods.minDate)"
...
```

Our Jackson [FieldFilterSimpleBeanPropertyFilter](https://github.com/dhis2/dhis2-core/blob/a7c9affd49bbb02db3b48c2597e377f098bfb779/dhis-2/dhis-services/dhis-service-field-filtering/src/main/java/org/hisp/dhis/fieldfiltering/FieldFilterSimpleBeanPropertyFilter.java#L118-L150) is then called for every field on an object to decide whether it should be serialized. [FieldFilterSimpleBeanPropertyFilter](https://github.com/dhis2/dhis2-core/blob/a7c9affd49bbb02db3b48c2597e377f098bfb779/dhis-2/dhis-services/dhis-service-field-filtering/src/main/java/org/hisp/dhis/fieldfiltering/FieldFilterSimpleBeanPropertyFilter.java#L118-L150) builds the full path again and again for every field from Jacksons generator state. This involves lots of String splits, joins, StringBuilder operations like `insert(0` 😨.

To answer the question `should the field be include it or not?` we then iterate over all field paths (which can really be all or most of them with `fields=*`) and [String compare paths](https://github.com/dhis2/dhis2-core/blob/a7c9affd49bbb02db3b48c2597e377f098bfb779/dhis-2/dhis-services/dhis-service-field-filtering/src/main/java/org/hisp/dhis/fieldfiltering/FieldFilterSimpleBeanPropertyFilter.java#L109-L110).

2. We are using a post-processing approach:

We serialize the [Jackson filtered object to a Jackson JsonNode](https://github.com/dhis2/dhis2-core/blob/4b34f59a62e1ebb2c28d1f4e9fee3497587d093c/dhis-2/dhis-services/dhis-service-field-filtering/src/main/java/org/hisp/dhis/fieldfiltering/FieldFilterService.java#L271) so we can then apply [transformations](https://docs.dhis2.org/en/develop/using-the-api/dhis-core-version-master/metadata.html#webapi_field_transformers)

https://github.com/FasterXML/jackson-databind/blob/d4220ed27198847e13ce75d9bee805e028a100e0/src/main/java/com/fasterxml/jackson/databind/ObjectMapper.java#L3639
> Functionally similar to serializing value into token stream and parsing that stream back as tree model node, but more efficient as {@link TokenBuffer} is used to contain the intermediate representation instead of fully serialized contents.

The price of field transformations is also paid when users don't actually transform any fields. Remember [valueToTree](https://github.com/FasterXML/jackson-databind/blob/d4220ed27198847e13ce75d9bee805e028a100e0/src/main/java/com/fasterxml/jackson/databind/ObjectMapper.java#L3636-L3642) described above.

This is the flow of data until it reaches the HTTP body

`DHIS2 entity => Jackson property filter & serialize to a Jackson ObjectNode => serialized again in controllers/message converter to HTTP body`

So even though you find the words `stream` in some of the involved code we do not actually stream the entities to the HTTP body directly. We will create JSON in-memory representations first like in [StreamingJsonRoot.pager](https://github.com/dhis2/dhis2-core/blob/4b34f59a62e1ebb2c28d1f4e9fee3497587d093c/dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/webdomain/StreamingJsonRoot.java#L47-L48) which is then serialized in [StreamingJsonRootMessageConverter](https://github.com/dhis2/dhis2-core/blob/4b34f59a62e1ebb2c28d1f4e9fee3497587d093c/dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/mvc/messageconverter/StreamingJsonRootMessageConverter.java#L62) when it comes to metadata. Endpoints returning a Jackson `ObjectNode` or any other type rely on Spring serializing to the HTTP body using the `ObjectMapper`.

3. (maybe minor) Not following [Jacksons advice](https://github.com/fasterxml/jackson-docs/wiki/presentation:-jackson-performance#basics-things-you-should-do-anyway) as we copy the heavy thread safe [ObjectMapper on every request](https://github.com/dhis2/dhis2-core/blob/4b34f59a62e1ebb2c28d1f4e9fee3497587d093c/dhis-2/dhis-services/dhis-service-field-filtering/src/main/java/org/hisp/dhis/fieldfiltering/FieldFilterService.java#L258).

It is important we optimize this
* as it affects almost every endpoint
* even when users don't send us a `fields` parameter a default will be in place like this one in `/tracker/trackedEntities` [fields=*,!relationships,!enrollments,!events,!programOwners](https://github.com/dhis2/dhis2-core/blob/4b34f59a62e1ebb2c28d1f4e9fee3497587d093c/dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/tracker/export/trackedentity/TrackedEntityRequestParams.java#L68)
* you could argue users rarely use a `pageSize` larger than the default of `50`. It still matters as we have deeply nested entities like tracked entity that can have lots of collections of nested entities which are not affected by the `pageSize`
* our apps like Capture make many API requests even in a single page. Try clicking your way to an event dashboard by selecting an orgUnit and program and you'll find ~80 API requests (the exact number of API requests is lower due to some being for static content or not field filtered)

## Improved implementation

This PR shows how we can improve what was [re-implemented in 2.38](https://dhis2.atlassian.net/browse/TECH-650) using Jackson. The essence is that for every field Jackson needs to know `should I include it or not?`. This is essentially a predicate `boolean test(String field)` which must be answered quickly.

How?

1. We build up a structure of predicates (`Fields`) that reflect the structure of the data being filtered. [in `FieldsParser`]
2. We use the fact that Jackson will traverse the object tree in depth-first order. [in `FieldsPropertyFilter`]
3. We store the current objects predicate tree in Jacksons context and reset it once Jackson is done with serializing that object. [in `FieldsPropertyFilter`]
6. We do still use the current approach of building up the intermediate `Jackson ObjectNode` but only for fields that are actually transformed. [in `FieldsPropertyFilter`]

Transformations (4.) could be further improved. Some transformations like `rename` can be done without creating intermediate `Jackson ObjectNode`. Others like `isEmpty` are more challenging as they need JSON type information which we do not have in the property filter unless we actually serialize. I did not go further with optimizing 4. as I am not sure many people even use transformations.

Benefits of new implementation

* faster response times
* less garbage as without transformation we truly stream our Java entities to the HTTP body

## Benchmark

Benchmark `FieldFilterSerializationBenchmarkTest` compares the throughput of the current and the proposed implementation. It also compares the presence of field filtering when no filtering occurs i.e. `fields=*` with pure Jackson serialization. The **operation** being benchmarked is the parsing of `fields` and the serialization of `List<Event>` to a JSON `String` which includes filtering and transforming events.

JMH advice is to fork the JVM to prevent interference between the runs. I have set `fork(0)` because of Spring/authentication. It took me a while to get Spring and JMH to work. So don't put too much faith into the exact numbers (I also ran this locally). Since our workload is static, same POJOs are serialized, I think that should be alright. The difference in magnitude should also be enough indication. I also ran a performance test locally against the application so we can more clearly see the impact on the response times.

### `fields=*` (overhead analysis)

We can calculate an overhead of the `field filter` feature/implementation by comparing the field filter implementations throughput with `fields=*` to pure Jackson serialization.

<img width="3550" height="2364" alt="benchmark_graph_star" src="https://github.com/user-attachments/assets/63dc8a88-0611-4789-8071-370f695118ec" />

These are the throughput values and the overhead i.e. `Overhead = (noFieldFiltering - xxxFieldFiltering) / noFieldFiltering`. The event count is the number of events serialized in the benchmarked operation.

|   Event Count | No Filtering (events/s)   | Better Filtering (events/s)   | Current Filtering (events/s)   | Better Overhead (%)   | Current Overhead (%)   | Better vs Current (x faster)   |
|--------------:|:--------------------------|:------------------------------|:-------------------------------|:----------------------|:-----------------------|:-------------------------------|
|            25 | 104,891                   | 63,939                        | 1,446                          | 39.0%                 | 98.6%                  | 44.2x                          |
|            50 | 84,331                    | 57,366                        | 1,504                          | 32.0%                 | 98.2%                  | 38.1x                          |
|           100 | 88,863                    | 53,817                        | 1,597                          | 39.4%                 | 98.2%                  | 33.7x                          |
|           200 | 91,592                    | 51,846                        | 1,637                          | 43.4%                 | 98.2%                  | 31.7x                          |
|           400 | 88,885                    | 55,407                        | 1,599                          | 37.7%                 | 98.2%                  | 34.6x                          |
|           800 | 91,857                    | 53,514                        | 1,694                          | 41.7%                 | 98.2%                  | 31.6x                          |

#### Allocations and GC

I ran the `fields=*` and `eventCount=50` benchmark from before with the async-profiler profiling allocations. In this run the proposed implementation processed 31x more events/s than the current implementation.

|   Event Count | No Filtering (events/s)   | Better Filtering (events/s)   | Current Filtering (events/s)   | Better Overhead (%)   | Current Overhead (%)   | Better vs Current (x faster)   |
|--------------:|:--------------------------|:------------------------------|:-------------------------------|:----------------------|:-----------------------|:-------------------------------|
|            50 | 71,569                    | 40,885                        | 1,333                          | 42.9%                 | 98.1%                  | 30.7x                          |

This was the allocation behavior:

  | Event Type                      | No Filtering   | Better         | Current       |
  |---------------------------------|----------------|----------------|---------------|
  | jdk.ObjectAllocationInNewTLAB   | 29,468 (592KB) | 16,756 (333KB) | 6,309 (120KB) |
  | jdk.ObjectAllocationOutsideTLAB | 0              | 0              | 0             |
  | jdk.GCHeapSummary               | 24 events      | 14 events      | 6 events      |

if we now compare the allocations per event

| Implementation | Total Allocs | Events/s | Duration | Total Events | Allocs/Event |
  |----------------|--------------|----------|----------|--------------|--------------|
  | Current        | 6,309        | 1,333    | 25s      | 33,325       | 0.189        |
  | Better         | 16,756       | 40,885   | 25s      | 1,022,125    | 0.016        |
  | No Filtering   | 29,468       | 71,569   | 25s      | 1,789,225    | 0.016        |

- Better implementation: 12x more allocation-efficient per event (0.016 vs 0.189)
- Better optimization: Achieves baseline allocation efficiency (same as no filtering) while maintaining filtering functionality
- The current implementation has less GC events which I would explain by it processing considerably less events

Allocation flamegraphs created using `jfrconv --alloc --total jfr-alloc.jfr alloc-flamegraph.html`:

- no field filtering [alloc-flamegraph.html](https://github.com/user-attachments/files/22166981/alloc-flamegraph.html)
- better field filtering [alloc-flamegraph.html](https://github.com/user-attachments/files/22166995/alloc-flamegraph.html)
- current field filtering [alloc-flamegraph.html](https://github.com/user-attachments/files/22167000/alloc-flamegraph.html)

You can see the current field filtering does lots of `String` manipulations. The better implementation shows allocations due to Jackson and date serialization which we can optimize further later on.

### `fields=*,!relationships`

|   Event Count | Better Filtering (events/s)   | Current Filtering (events/s)   | Performance Improvement   |
|--------------:|:------------------------------|:-------------------------------|:--------------------------|
|            25 | 188,110                       | 14,115                         | 13.3x                     |
|            50 | 185,000                       | 21,462                         | 8.6x                      |
|           100 | 172,609                       | 26,726                         | 6.5x                      |
|           200 | 165,711                       | 31,444                         | 5.3x                      |
|           400 | 171,590                       | 34,400                         | 5.0x                      |
|           800 | 165,264                       | 35,154                         | 4.7x                      |

### `fields=*,event~rename(foo)` (transformation)

From the results of serializing all fields with one renamed we can see

* The current implementation has similar throughput to `fields=*` as it always creates the intermediate `Jackson ObjectNode` in memory.
* You can see the transformation does affect the proposed implementation as it uses the same approach and creates an intermediate `Jackson ObjectNode` in memory. It still outperforms the current implementation as it will only create an `ObjectNode` for the node it actually transforms.

|   Event Count | Better Filtering (events/s)   | Current Filtering (events/s)   | Performance Improvement   |
|--------------:|:------------------------------|:-------------------------------|:--------------------------|
|            25 | 56,172                        | 1,461                          | 38.4x                     |
|            50 | 50,329                        | 1,521                          | 33.1x                     |
|           100 | 53,232                        | 1,657                          | 32.1x                     |
|           200 | 50,400                        | 1,584                          | 31.8x                     |
|           400 | 53,804                        | 1,588                          | 33.9x                     |
|           800 | 52,718                        | 1,659                          | 31.8x                     |

### `fields=event`

Filtering out all but a single string field with `fields=event` leads to

|   Event Count | Better Filtering (events/s)   | Current Filtering (events/s)   | Performance Improvement   |
|--------------:|:------------------------------|:-------------------------------|:--------------------------|
|            25 | 2,146,797                     | 95,280                         | 22.5x                     |
|            50 | 2,220,312                     | 162,887                        | 13.6x                     |
|           100 | 2,356,845                     | 248,318                        | 9.5x                      |
|           200 | 2,447,940                     | 343,439                        | 7.1x                      |
|           400 | 2,274,336                     | 416,872                        | 5.5x                      |
|           800 | 2,479,009                     | 445,075                        | 5.6x                      |

## HTTP API performance

I ran a [Gatling](https://github.com/dhis2/performance-tests-gatling) performance tests locally with a single user making 100 sequential HTTP requests against

1. dhis2/core-dev:42-47f0d4eae9 with no code changes
2. this PRs changes with the proposed implementation

I picked a slow request the Capture app makes as the target. Select a [Child Programme and orgUnit Ngelehun CHC](https://play.im.dhis2.org/stable-2-42-1/apps/capture#/?orgUnitId=DiszpKrYNg8&programId=IpHINAT79UW&selectedTemplateId=IpHINAT79UW-default). Capture app then makes request `https://play.im.dhis2.org/stable-2-42-1/api/42/tracker/trackedEntities?order=createdAt:desc&page=1&pageSize=15&orgUnits=DiszpKrYNg8&orgUnitMode=SELECTED&program=IpHINAT79UW&fields=:all,!relationships,programOwners[orgUnit,program]`.

The default `pageSize` on the Capture app page above is `15` with options `10, 15, 25, 50, 100`. This graph shows the difference in response times of `current - proposed`

<img width="1000" height="600" alt="analysis" src="https://github.com/user-attachments/assets/580aab39-6cee-4b3a-83e0-feb13a887a00" />

…